Reduce refresh bars time, clean code

+++ src/main/java/com/arturoo404/game/utils/StatsUtils.java
@@ -2,7 +2,7 @@ package com.arturoo404.game.utils;
 
 public class StatsUtils {
 
-    public static int calcDamageAfterReduction(int entityDamage, int armor){
-        return entityDamage * (1 - (armor / (armor + 100)));
+    public static int calcDamageAfterReduction(double entityDamage, double armor){
+        return (int) (entityDamage * (1 - (armor / (armor + 100))));
     }
 }
